Code cleanup

+++ code/src/io/JewelStripe.h
@@ -3,8 +3,7 @@
 
 #include "Jewel.h"
 
-class JewelStripe
-{
+class JewelStripe {
 
 public:
 
@@ -13,8 +12,7 @@ public:
      * @param pin The pin number on which the NeoPixel is connected.
      * @param numberOfJewels The number of Jewels composing the stripe.
      */
-    JewelStripe(uint8_t pin, uint8_t numberOfJewels)
-    {
+    JewelStripe(uint8_t pin, uint8_t numberOfJewels) {
         uint8_t numberOfPixels = Jewel::NUMBER_OF_PIXELS_PER_JEWEL * numberOfJewels;
         _stripe = new Jewel(pin, numberOfPixels);
     }
@@ -28,8 +26,7 @@ public:
      * @param pixelNumber The pixel to apply the color.
      * @param pixelColor The color represented in hexadecimal.
      */
-    void setPixelColor(uint8_t pixelNumber, uint64_t pixelColor)
-    {
+    void setPixelColor(uint8_t pixelNumber, uint32_t pixelColor) {
         _stripe->setPixelColor(pixelNumber, pixelColor);
     }
 
@@ -41,8 +38,7 @@ public:
      * @param blue The intensity of blue for the color (between 0 and 255).
      * @param white The intensity of white for the color (between 0 and 255).
      */
-    void setPixelColor(uint8_t pixelNumber, uint8_t red, uint8_t green,  uint8_t blue, uint8_t white)
-    {
+    void setPixelColor(uint8_t pixelNumber, uint8_t red, uint8_t green,  uint8_t blue, uint8_t white) {
         _stripe->setPixelColor(pixelNumber, red, green, blue, white);
     }
 
@@ -50,8 +46,7 @@ public:
      * Set the colors for all pixels.
      * @param pixelColor The color represented in hexadecimal.
      */
-    void setPixelsColor(uint64_t pixelColor)
-    {
+    void setPixelsColor(uint32_t pixelColor) {
         _stripe->setPixelsColor(pixelColor);
     }
 
@@ -62,21 +57,18 @@ public:
      * @param blue The intensity of blue for the color (between 0 and 255).
      * @param white The intensity of white for the color (between 0 and 255).
      */
-    void setPixelsColor(uint8_t red, uint8_t green,  uint8_t blue, uint8_t white)
-    {
+    void setPixelsColor(uint8_t red, uint8_t green,  uint8_t blue, uint8_t white) {
         _stripe->setPixelsColor(red, green, blue, white);
     }
 
     /**
      * Return the number of pixels.
      */
-    uint8_t numberOfPixels()
-    {
+    uint8_t numberOfPixels() {
       return _stripe->numberOfPixels();
     }
 
 protected:
-
     //! Stripe of jewels
     Jewel* _stripe;
 };
